fix: lazy fix 1.20.1 forge through sinytra by disabling some mixins
