+++ common/src/main/java/de/zannagh/armorhider/CompatFlags.java
@@ -13,6 +13,7 @@ public final class CompatFlags {
     public static boolean ET_LOADED = false;
     public static boolean FA_LOADED = false;
     public static boolean GECKOLIB_LOADED = false;
+    public static boolean SINYTRA_LOADED = false;
     public static boolean WFGM_LOADED = false;
 
     private CompatFlags() {}
